feat: complete config rewrite
The new configuration is using a custom function to create a new
configuration. The function is adapter for specifically my configuration
and may need to be modified at a later date. This need to be fixed.

Todo: nix modules for other features outisde home-managers capabilities

The system configuration is stored within modules which can be enabled
at will using host specific configuration files.

Currently three configuration files per host to keep track of.

`user.nix` will always be present since home-manager and the modularized
features is configured using that file.

`hardware-configuration.nix` is the file that NEED to be generated using
the host machine and will be specific to that machine and that machine
only. **IMPORTANT**

`disk-config.nix` is the file where disko configuration is stored. This
file will only be included when disco is enabled.